ordene un poco el codigo, restructure las carpetas del proyecto y mejore el README.md
