+++ README.MD
@@ -1,0 +1,20 @@
+Hola, este es el programa de los alumnos: 
+- Marcos Gabriel Pereyra
+- Leonardo David Paez
+- Miguel Angel Luis Rubin Azas 
+- Marcelo Ruiz Burgos.
+
+Este programa se encarga de Tomar asistencias de los alumnos con tan solo una cámara. Solo deben ponerse en frente de ella para que, si su rostro está registrado, lo anote en una lista de alumnos precentes.
+
+Este programa cuenta con las librerías: ```OpenCV, Os, Numpy 1.19.3, imutils```. (estas librerías se podrán instlalar con el archivo ```pip_archivo.bat``` en caso de necesitarlo)
+Tambien cuenta con el framework: ```haarcascade_frontalface_default.xml``` de Malayanil cuyo repositorio dejaré a continuación: 
+
+https://github.com/Malayanil/FaceDetectionHC
+
+debido a inconvenientes que aún no sabemos nuestro proyecto no puede ser comprimido como un .exe
+
+a continuación mostraré los pasos para abrir el proyecto:
+- Tener Cámara conectáda a la computadora
+- Tener instalado Python 3.9 en su Computadora
+- (Solo hace falta instalarlo una sola vez) Instalador de los Archivos necesarios para ejecutar "Facial Scan Assistance.bat" abriendo el archivo llamado "pip_archivo.bat"
+- Ejecuar el archivo llamado "Facial Scan Assistance.bat"
