+++ depedencias/pip_archivo.bat
@@ -1,0 +1,4 @@
+pip install pip
+pip install numpy==1.19.3
+pip install imutils
+pip install opencv-contrib-python==4.4.0.46
